Solution for Day 34

+++ Day34.cpp
@@ -2,8 +2,8 @@
 
 #include <iostream>
 
-long long main() {
-    int n;
+int main() {
+    long long n;
     std::cin >> n;
     if (n % 2 == 0) {
         std::cout << n/2 << std::endl;
